Allow ElasticsearchLogClient factory to be passed as option
The `ElasticsearchLogClient` uses internally `DynamicResponse` object that is not usable in when application is compiled to native code. This is usually the case with UWP apps as those are compiled to native code by default on release mode.

So this change allows `ElasticsearchLogClient` to injected to `DurableElasticsearchSink` so that the client can be overridden with custom implementation when the default `ElasticsearchLogClient` cannot be used.

For more information about the issue with `DynamicResponse` and native code compilation see elastic/elasticsearch-net#4389